// lib/enums/enums.ts
export enum IconEnum {
    DELETE = "DELETE",
    REMOVE = "REMOVE",
    ADD = "ADD",
    EDIT = "EDIT",
    SORT_ASC = "SORT_ACS",
    SORT_DESC = "SORT_DES",
    SORT_NEUTRAL = "SORT_NEUTRAL",
    DEACTIVE = "DEACTIVE",
    ACTIVE = "ACTIVE",
    ENABLED = "ENABLED",
    DISABLED = "DISABLED"
}

export enum DateFormatEnum {
    DISPLAY_DATE = "DD-MMM-YYYY",
    CUSTOM_DATE = 'DD MMM YYYY [@]HH:mm:ss A',
    DATE4 = "YYYYMMDD",
    DATE4Dash = "YYYY-MM-DD",
    DATE4Slash = "YYYY/MM/DD",
}

export enum PathRoute {
    MY_COURSE = "/courses",
}

export enum localStroageEnum {
    DATEPICK = 'datepicker',
    VEIWCOLUMN = 'view-column'
    
}

export enum PopupTypeEnum {
    // USER
    CREATE_USER = 'CREATE_USER',
    USER_DETAILS = 'USER_DETAILS',
    CREATE_PARTNER = 'CREATE_PARTNER',
    SUSPEND_USER = 'SUSPEND_USER',
    REACTIVATE_USER = 'REACTIVATE_USER',
    INACTIVATE_USER = 'INACTIVATE_USER',
    DELETE_USER = 'DELETE_USER',
    EDIT_USER = 'EDIT_USER',
    UPDATE_USER = 'UPDATE_USER',

    // PARTNER
    PARTNER_DETAILS = 'PARTNER_DETAILS',
    UPDATE_PARTNER = 'UPDATE_PARTNER',
    DELETE_PARTNER = 'DELETE_PARTNER',
    REACTIVATE_PARTNER = 'REACTIVATE_PARTNER',
    INACTIVATE_PARTNER = 'INACTIVATE_PARTNER',

    //API PERMISSION
    API_PERMISSION_INFO="API_PERMISSION_INFO",
    API_PERMISSION_EDIT="API_PERMISSION_EDIT",
    API_PERMISSION_ASSIGN="API_PERMISSION_ASSIGN",
    API_PERMISSION_DELETE="API_PERMISSION_DELETE",
    API_PERMISSION_STATUS="API_PERMISSION_STATUS",
    API_PERMISSION_ENABLED="API_PERMISSION_ENABLED",
    API_PERMISSION_DISABLE="API_PERMISSION_DISABLE",

    //API MANAGEMENT
    API_MANAGEMENT_REGISTER="API_MANAGEMENT_REGISTER",
    API_MANAGEMENT_VIEW="API_MANAGEMENT_VIEW",
    API_CATEGORY_CREATE="API_CATEGORY_CREATE",

    //AUTH CONFIG
    AUTH_CONFIG_INFO="AUTH_CONFIG_INFO",
    AUTH_CONFIG_EDIT="AUTH_CONFIG_EDIT",
    AUTH_CONFIG_CREATE="AUTH_CONFIG_CREATE",
    AUTH_CONFIG_DELETE="AUTH_CONFIG_DELETE",
    AUTH_CONFIG_STATUS="AUTH_CONFIG_STATUS",
    AUTH_CONFIG_ENABLED="AUTH_CONFIG_ENABLED",
    AUTH_CONFIG_DISABLE="AUTH_CONFIG_DISABLE",

    // COURSE
    COURSE_DETAILS = 'COURSE_DETAILS',

    // COMMON
    DELETE_CONFIRMATION = 'DELETE_CONFIRMATION',
    RESET_PASSOWRD = 'RESET_PASSWORD',
    VIEW_SETTINGS = 'VIEW_SETTINGS',
    LOG_OUT = "LOG_OUT",
    NONE = 'NONE',
}

export enum RowActionEnum {
    USER = 'USER',
    PARTNER = 'PARTNER',
    API_PERMISSION = 'API_PERMISSION',
    API_MANAGEMENT = 'API_MANAGEMENT',
    AUTH_CONFIG = 'AUTH_CONFIG',
}

export enum RoleEnum {
    USER = 'USER',
    ADMIN = 'ADMIN',
    PARTNER = 'PARTNER',
}

export enum UserStatusEnum {
    ACTIVE = '1',
    INACTIVE = '2'
}


export enum StatusEnum {
    ACTIVE = "Active",
    INACTIVE = "Inactive",
    ENABLED = "Enabled",
    DISABLED = "Disabled"
}

export enum CommonCode {
    AUTH_TYPE = 'AUTH_TYPE',
}

export enum BusinessType {
    RETAILER = "BUSN001",
    BUYER = "BUSN002",
}

export enum APIStatus {
    ENABLED = "1",
    DISABLED = "2",
    DELETE = "9"
}